Move json into its own file so it can be used as a library
