+++ json_eval.cpp
@@ -1,4 +1,4 @@
-#include "json_eval.h"
+#include "json.h"
 
 
 int main(int argc, char** argv) {
@@ -16,18 +16,4 @@ int main(int argc, char** argv) {
     //JsonFile.close();
 
     return 0;
-}
-
-std::shared_ptr<Json::Node> Json::readJson(const std::string &filename) {
-    std::ifstream jsonFile;
-    jsonFile.open(filename);
-    std::string line;
-    //std::cout << filename << " is open: " << jsonFile.is_open() << std::endl;
-    if (jsonFile.is_open()) {
-        while (std::getline(jsonFile, line)) {
-            std::cout << " - " << line << std::endl;
-        }
-    } 
-    jsonFile.close();
-    return std::make_shared<Json::Value>("test", "b");
 }
